gen-ai/orchestrator-server/src/main/python/server/pyproject.toml
[tool.poetry]
name = "gen_ai_orchestrator"
version = "0.1.0"
description = ""
authors = ["Mohamed ASSOUKTI <mohamed.assoukti@partnre.com>"]
readme = "README.md"
packages = [{include = "gen_ai_orchestrator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.31.1"
pydantic-settings="^2.5.2"
fastapi = "^0.115.0"
langchain = "^0.3.3"
langchain-community = "^0.3.2"
langchain-openai = "^0.2.2"
tiktoken = "^0.8.0"
opensearch-py = "^2.7.1"
path = "^17.0.0"
colorlog = "^6.8.2"
boto3 = "^1.35.37"
urllib3 = "^2.2.3"
jinja2 = "^3.1.4"
langfuse = "^2.52.0"
httpx-auth-awssigv4 = "^0.1.4"
langchain-postgres = "^0.0.12"
google-cloud-secret-manager = "^2.20.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pip-audit = "^2.7.0"
bandit = {version = "^1.7.7", extras = ["json"]}
cyclonedx-bom = "^4.1.4"

[tool.poetry.group.test.dependencies]
tox = "^4.11.4"
coverage = "^7.4.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.6"

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"